requirements.txt
gunicorn
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==36.0.0
distlib==0.3.3
Django==3.2.9
djangorestframework==3.12.4
filelock==3.0.12
h11==0.12.0
idna==3.3
mouse==0.7.1
outcome==1.1.0
Pillow==8.4.0
platformdirs==2.3.0
pycparser==2.21
pyOpenSSL==21.0.0
pytz==2021.3
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3
sqlparse==0.4.2
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
virtualenv==20.8.1
wsproto==1.0.0